Refactor signature generation for simplicity (#8572)
